Firing fully implemented
However, the game is starting to get a little laggy, which I was really hoping wouldn't be a problem. I believe it is the upgrades, as occasionally there will be a slight spike right after a choice. First I'll put in the rest of the upgrades, then let's see if we can fix that. I think I just spent 6+ hours on putting firing (with bouncing) in without bugs.