Major commit
This can be considered a new build, many changes were made.
Art: Went through several iterations of the food for the snake, worked for about 15 minutes on the shuttle home on some, I liked the acorn best and am using it but the grape is the TempFood currently.
Lasers: Once again rebalanced, this is a little more in line with what I think should work for the demo. Also brought them up to date with save data.
Reverse: Fully implemented in the way that I think is actually fun to use now. Remove a "MoveSnake()" in the method for a smoother bounce but some clipping, tradeoff not easily fixed.
Health: Implemented this which works EXTREMELY well with reverse. Probably my favorite feature right now with extra food coming second and the simplicity of dash third. the input queue is fourth, but I'm not gonna keep going on why my game is so fun and the best.
Timer: Added a fully functioning run timer that kills you at the end. This will be one of the main features and I believe is the penultimate piece in this game's framework.
Main Menu: implemented upgrade system there fully but need to work on meta currency after getting all upgrades implemented.
Misc: camera smoothing, tried and failed to make segments get hurt too, may require health becoming a float (health bar!!!)
Overall:
Spent maybe 5-7 hours on the game today, hope to log more in the next few days of break!