All Audio implemented correctly
Next on the todo is remove the unneeded Audio Player scripts and maybe make the scene switch faster if possible. Also, you can glitch out of the map by going up to a wall and hitting against it until you break through. This seems to be ingrained in the movement system (reverse and tick speed are involved) so I'm not sure if I'll end up fixing it.